Update and rename Auto Clicker to Auto Clicker.py
You can add input validation to make sure that the interval entered by the user is a positive integer. You can also add a default value for the interval to use if the user does not enter a valid value.

Instead of using an infinite loop and a break statement to stop the bot, you can use a flag variable to control the loop. This can make the code easier to understand and maintain.

Instead of using time.sleep to pause the program before starting the bot, you can use a timer or a countdown label to give the user a visual indication of when the bot will start.

You can add a stop button to allow the user to stop the bot at any time.

You can add error handling to catch and handle any exceptions that might occur while running the bot.

You can use more descriptive names for the functions and variables. For example, instead of clickedStart, you could use start_bot.

Here is an example of how the improved code could look: